Create stub test for relevant transform. Correct test file path. Reenable schema validation tests (now that we've clipped the obsolete SARIF reference).
