+++ src/Sarif.UnitTests/TestData/PrereleaseCompatibilityTransformer/ExpectedOutputs/RunResources.sarif
@@ -1,0 +1,63 @@
+{
+  "$schema": "http://json.schemastore.org/sarif-2.0.0-csd.2.beta.2018-11-28",
+  "version": "2.0.0-csd.2.beta.2018-11-28",
+  "runs": [
+    {
+      "tool": {
+        "name": "CodeScanner",
+        "semanticVersion": "2.1.0"
+      },
+      "results": [
+        {
+          "ruleId": "SHARED-ID",
+          "ruleIndex": 1,
+          "message": {
+            "messageId": "a"
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "fileLocation": {
+                  "uri": "file.txt"
+                }
+              }
+            }
+          ]
+        },
+        {
+          "ruleId": "SHARED-ID",
+          "ruleIndex": 0,
+          "message": {
+            "messageId": "a"
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "fileLocation": {
+                  "uri": "file.png"
+                }
+              }
+            }
+          ]
+        }
+      ],
+      "resources": {
+        "rules": [
+          {
+            "id": "SHARED-ID",
+            "messageStrings": {
+              "a": "Review all image content for geopolitically sensitive graphics."
+            }
+          },
+          {
+            "id": "SHARED-ID",
+            "messageStrings": {
+              "a": "Review all plaintext content for geopolitically sensitive terms."
+            }
+          }
+        ]
+      },
+      "columnKind": "utf16CodeUnits"
+    }
+  ]
+}
