+++ src/Sarif.UnitTests/TestData/PrereleaseCompatibilityTransformer/Inputs/RunResources.sarif
@@ -1,0 +1,36 @@
+{
+  "$schema": "http://json.schemastore.org/sarif-2.0.0-csd.2.beta.2018-10-10",
+  "version": "2.0.0-csd.2.beta.2018-10-10",
+  "runs": [
+    {
+      "tool": {
+        "name": "CodeScanner",
+        "semanticVersion": "2.1.0"
+      },
+      "resources": {
+        "rules": {
+          "SHARED-ID.2": {
+            "id": "SHARED-ID",
+            "messageStrings": { "a": "Review all image content for geopolitically sensitive graphics." }
+          },
+          "SHARED-ID.1": {
+            "id": "SHARED-ID",
+            "messageStrings": { "a": "Review all plaintext content for geopolitically sensitive terms." }
+          }
+        }
+      },
+      "results": [
+        {
+          "ruleId": "SHARED-ID.1",
+          "message": { "messageId": "a" },
+          "locations": [ { "physicalLocation": { "fileLocation": { "uri": "file.txt" } } } ]
+        },
+        {
+          "ruleId": "SHARED-ID.2",
+          "message": { "messageId": "a" },
+          "locations": [ { "physicalLocation": { "fileLocation": { "uri": "file.png" } } } ]
+        }
+      ]
+    }
+  ]
+}
